requirements.txt
# Core numerics
numpy>=1.26
scipy>=1.12
pandas>=2.2

# Seismology + ML (shared with IDA-PTW)
obspy>=1.4
seisbench>=0.6
torch>=2.3
xgboost>=2.1
scikit-learn>=1.5

# MQTT
paho-mqtt>=2.1
gmqtt>=0.7            # async MQTT5 client for inference services

# Config + schemas
pyyaml>=6.0
pydantic>=2.7

# Logging / obs
structlog>=24.1
python-json-logger>=2.0

# Dev / test
pytest>=8.2
pytest-asyncio>=0.23
ruff>=0.5
